chore: refactored the entry point and added comments
